+++ DependencyInjection/Configuration.php
@@ -15,6 +15,7 @@ use Dmishh\SettingsBundle\Manager\SettingsManagerInterface;
 use Symfony\Component\Config\Definition\Builder\TreeBuilder;
 use Symfony\Component\Config\Definition\ConfigurationInterface;
 use Symfony\Component\Config\Definition\Exception\InvalidTypeException;
+use Symfony\Component\Form\Extension\Core\Type\TextType;
 
 class Configuration implements ConfigurationInterface
 {
@@ -23,14 +24,19 @@ class Configuration implements ConfigurationInterface
      */
     public function getConfigTreeBuilder()
     {
-        $treeBuilder = new TreeBuilder();
-        $rootNode = $treeBuilder->root('dmishh_settings');
+        $treeBuilder = new TreeBuilder('dmishh_settings');
+        // Keep compatibility with symfony/config < 4.2
+        if (!method_exists($treeBuilder, 'getRootNode')) {
+            $rootNode = $treeBuilder->root('dmishh_settings');
+        } else {
+            $rootNode = $treeBuilder->getRootNode();
+        }
 
-        $scopes = array(
+        $scopes = [
             SettingsManagerInterface::SCOPE_ALL,
             SettingsManagerInterface::SCOPE_GLOBAL,
             SettingsManagerInterface::SCOPE_USER,
-        );
+        ];
 
         $rootNode
             ->children()
@@ -42,13 +48,13 @@ class Configuration implements ConfigurationInterface
                 ->arrayNode('security')
                     ->addDefaultsIfNotSet()
                     ->children()
-                        ->scalarNode('manage_global_settings_role')->end()
+                        ->scalarNode('manage_global_settings_role')->defaultValue(null)->end()
                         ->booleanNode('users_can_manage_own_settings')->defaultValue(true)->end()
                     ->end()
                 ->end()
                 ->enumNode('serialization')
                     ->defaultValue('php')
-                    ->values(array('php', 'json'))
+                    ->values(['php', 'json'])
                 ->end()
                 ->arrayNode('settings')
                     ->prototype('array')
@@ -61,14 +67,14 @@ class Configuration implements ConfigurationInterface
                                     ->thenInvalid('Invalid scope %s. Valid scopes are: '.implode(', ', array_map(function ($s) { return '"'.$s.'"'; }, $scopes)).'.')
                                 ->end()
                             ->end()
-                            ->scalarNode('type')->defaultValue('Symfony\Component\Form\Extension\Core\Type\TextType')->end()
+                            ->scalarNode('type')->defaultValue(TextType::class)->end()
 
                             ->variableNode('options')
                                 ->info('The options given to the form builder')
-                                ->defaultValue(array())
+                                ->defaultValue([])
                                 ->validate()
                                     ->always(function ($v) {
-                                        if (!is_array($v)) {
+                                        if (!\is_array($v)) {
                                             throw new InvalidTypeException();
                                         }
 
@@ -78,10 +84,10 @@ class Configuration implements ConfigurationInterface
                             ->end()
                             ->variableNode('constraints')
                                 ->info('The constraints on this option. Example, use constraits found in Symfony\Component\Validator\Constraints')
-                                ->defaultValue(array())
+                                ->defaultValue([])
                                 ->validate()
                                     ->always(function ($v) {
-                                        if (!is_array($v)) {
+                                        if (!\is_array($v)) {
                                             throw new InvalidTypeException();
                                         }