+++ Exception/WrongScopeException.php
@@ -13,17 +13,17 @@ namespace Dmishh\SettingsBundle\Exception;
 
 use Dmishh\SettingsBundle\Manager\SettingsManagerInterface;
 
-class WrongScopeException extends SettingsException
+class WrongScopeException extends \LogicException implements SettingsException
 {
     public function __construct($scope, $settingName)
     {
-        if ($scope === SettingsManagerInterface::SCOPE_GLOBAL) {
+        if (SettingsManagerInterface::SCOPE_GLOBAL === $scope) {
             $message = sprintf(
                 'You tried to access setting "%s" but it is in the "%s" scope which means you must not use a SettingOwnerInterface object with this option.',
                 $settingName,
                 $scope
             );
-        } elseif ($scope === SettingsManagerInterface::SCOPE_USER) {
+        } elseif (SettingsManagerInterface::SCOPE_USER === $scope) {
             $message = sprintf(
                 'You tried to access setting "%s" but it is in the "%s" scope which means you have to pass a SettingOwnerInterface object with this option.',
                 $settingName,