+++ Tests/Functional/ServiceInstantiationTest.php
@@ -2,22 +2,42 @@
 
 namespace Dmishh\SettingsBundle\Tests\Functional;
 
+use Dmishh\SettingsBundle\DmishhSettingsBundle;
 use Dmishh\SettingsBundle\Serializer\PhpSerializer;
-use Symfony\Bundle\FrameworkBundle\Test\WebTestCase;
+use Dmishh\SettingsBundle\Serializer\SerializerInterface;
+use Doctrine\Bundle\DoctrineBundle\DoctrineBundle;
+use Nyholm\BundleTest\BaseBundleTestCase;
+use Nyholm\BundleTest\CompilerPass\PublicServicePass;
 
-class ServiceInstantiationTest extends WebTestCase
+/**
+ * @internal
+ */
+final class ServiceInstantiationTest extends BaseBundleTestCase
 {
-    public static function setUpBeforeClass()
+    protected function setUp(): void
     {
-        static::bootKernel();
+        parent::setUp();
+
+        // Make services public that have an idea that matches a regex
+        $this->addCompilerPass(new PublicServicePass('|Dmishh.*|'));
     }
 
-    public function testHttpClient()
+    public function testInitBundle()
     {
-        $container = static::$kernel->getContainer();
+        $kernel = $this->createKernel();
+        $kernel->addConfigFile(\dirname(__DIR__).'/Resources/app/config/config.yml');
+        $kernel->addBundle(DoctrineBundle::class);
+        $this->bootKernel();
+        $container = $this->getContainer();
 
-        $this->assertTrue($container->has('dmishh.settings.serializer'));
-        $serializer = $container->get('dmishh.settings.serializer');
-        $this->assertInstanceOf(PhpSerializer::class, $serializer);
+        // Test if you services exists
+        self::assertTrue($container->has(SerializerInterface::class));
+        $service = $container->get(SerializerInterface::class);
+        self::assertInstanceOf(PhpSerializer::class, $service);
+    }
+
+    protected function getBundleClass()
+    {
+        return DmishhSettingsBundle::class;
     }
 }